{"schema":"decodex/commit/1","summary":"Modularize overlay session contracts","authority":"manual"}
